requirements-dev.txt
# Include base requirements
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
factory-boy>=3.2.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
pylint>=2.16.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Development tools
ipython>=8.0.0
pre-commit>=3.0.0
